requirements.txt
Flask>=2.1
python-dotenv>=0.21
SQLAlchemy>=2.0
Werkzeug>=2.0
PyJWT>=2.6
gunicorn
groq
